switch timeout millis from float to int

+++ opentelemetry-sdk/src/opentelemetry/sdk/_logs/_internal/export/__init__.py
@@ -122,7 +122,7 @@ class LogRecordExporter(abc.ABC):
         """
 
     @abc.abstractmethod
-    def force_flush(self, timeout_millis: float = 10_000) -> bool:
+    def force_flush(self, timeout_millis: int = 10_000) -> bool:
         """Hint to ensure that the export of any ``ReadableLogRecord`` objects
         the exporter has received prior to the call to ``force_flush`` SHOULD be
         completed as soon as possible, preferably before returning from this method.
@@ -171,7 +171,7 @@ class ConsoleLogRecordExporter(LogRecordExporter):
     def shutdown(self):
         pass
 
-    def force_flush(self, timeout_millis: float = 10_000) -> bool:
+    def force_flush(self, timeout_millis: int = 10_000) -> bool:
         return True
